Add initial implementation
This is the original prototype implementation. The code has been cleaned
up. Some features and most of the documentation are missing. Support for
wolfCrypt has been removed because wolfCrypt does not support GMAC with
multiple input chunks. Support for the legacy STM32 cryptographic
library has been removed in favor of the new STM32 crypto library (v4),
which is the Cortex-M Optimized Crypto Stack (CMOX).

All tests are expected to pass on Linux for the gcrypt, mbedtls, nettle,
and openssl backends. The cmox backend has been partially tested on a
Cortex-M4 MCU (STM32L432KC).