Fix design flaw in GMAC input construction
In auth-only mode, the original design used PKCS#7 padding for the AAD
and then appended the body when constructing the GMAC input. However,
this is clearly not safe, because multiple inputs lead to the same GMAC
input, which is highly undesirable.

The updated design fixes this issue by choosing a different
construction: the first block encodes the length of the AAD, followed by
the AAD itself and padding such that the AAD ends at a block boundary,
followed by the body (without padding). This construction is similar to
that used by ChaCha20-Poly1305, but less efficient because AES-GCM
itself will also pad the input and encode its length after we have
already done so.