Add CUBICRYPT_STM32_CMOX_AES_CRC0_WORKAROUND
This commit adds a new build option that, when enabled, causes internal
post-processing of primary keys and session keys that simulates the
behavior of CMOX (X-CUBE-CRYPTOLIB) when the hardware CRC unit is
disabled. This is for compatibility with a specific mission only and is
not to be enabled otherwise.

See https://gist.github.com/tniessen/cc8d0a86d830f5633098a0701021c8ab.